Inline DaycareService into DaycareController and CitizenDocumentResponseReport
