+++ service/src/main/kotlin/evaka/core/daycare/CaretakerQueries.kt
@@ -4,7 +4,6 @@
 
 package evaka.core.daycare
 
-import evaka.core.daycare.service.Caretakers
 import evaka.core.occupancy.GroupPredicate
 import evaka.core.shared.DaycareId
 import evaka.core.shared.GroupId
@@ -13,6 +12,18 @@ import evaka.core.shared.db.Database
 import evaka.core.shared.domain.FiniteDateRange
 import java.math.BigDecimal
 import java.time.LocalDate
+
+data class Caretakers(val minimum: Double, val maximum: Double) {
+    companion object {
+        fun fromDailyCounts(caretakers: DateMap<BigDecimal>): Caretakers =
+            Caretakers(
+                minimum =
+                    caretakers.entries().minOfOrNull { (_, count) -> count }?.toDouble() ?: 0.0,
+                maximum =
+                    caretakers.entries().maxOfOrNull { (_, count) -> count }?.toDouble() ?: 0.0,
+            )
+    }
+}
 
 fun Database.Transaction.initCaretakers(groupId: GroupId, startDate: LocalDate, amount: Double) {
     execute {
