+++ service/src/main/kotlin/evaka/core/daycare/controllers/DaycareController.kt
@@ -12,13 +12,17 @@ import evaka.core.application.getActiveTransferApplicationsFromUnit
 import evaka.core.backupcare.UnitBackupCare
 import evaka.core.backupcare.getBackupCaresForDaycare
 import evaka.core.daycare.CaretakerAmount
+import evaka.core.daycare.Caretakers
 import evaka.core.daycare.Daycare
 import evaka.core.daycare.DaycareFields
+import evaka.core.daycare.DaycareGroup
 import evaka.core.daycare.UnitFeatures
 import evaka.core.daycare.UnitOperationPeriod
 import evaka.core.daycare.addUnitFeatures
 import evaka.core.daycare.createDaycare
+import evaka.core.daycare.createDaycareGroup
 import evaka.core.daycare.deleteCaretakers
+import evaka.core.daycare.deleteDaycareGroup
 import evaka.core.daycare.getCaretakers
 import evaka.core.daycare.getDaycare
 import evaka.core.daycare.getDaycareGroup
@@ -33,16 +37,17 @@ import evaka.core.daycare.getLastPlacementDate
 import evaka.core.daycare.getOphUnitOIDs
 import evaka.core.daycare.getUnitFeatures
 import evaka.core.daycare.getUnitOperationPeriods
+import evaka.core.daycare.initCaretakers
 import evaka.core.daycare.insertCaretakers
+import evaka.core.daycare.isValidDaycareId
 import evaka.core.daycare.removeUnitFeatures
-import evaka.core.daycare.service.Caretakers
-import evaka.core.daycare.service.DaycareGroup
-import evaka.core.daycare.service.DaycareService
 import evaka.core.daycare.updateCaretakers
 import evaka.core.daycare.updateDaycare
 import evaka.core.daycare.updateGroup
 import evaka.core.daycare.updateUnitClosingDate
 import evaka.core.daycare.validateUnitClosingDate
+import evaka.core.messaging.createDaycareGroupMessageAccount
+import evaka.core.messaging.deleteDaycareGroupMessageAccount
 import evaka.core.occupancy.OccupancyPeriod
 import evaka.core.occupancy.OccupancyPeriodGroupLevel
 import evaka.core.occupancy.OccupancyResponse
@@ -58,6 +63,7 @@ import evaka.core.placement.getMissingGroupPlacements
 import evaka.core.placement.getTerminatedPlacements
 import evaka.core.placement.getUnitChildrenCapacities
 import evaka.core.placement.getWaitingUnitConfirmationApplicationsCount
+import evaka.core.placement.hasGroupPlacements
 import evaka.core.serviceneed.application.getUndecidedServiceApplicationsByUnit
 import evaka.core.shared.BackupCareId
 import evaka.core.shared.ChildId
@@ -68,7 +74,9 @@ import evaka.core.shared.GroupPlacementId
 import evaka.core.shared.PlacementId
 import evaka.core.shared.auth.AuthenticatedUser
 import evaka.core.shared.db.Database
+import evaka.core.shared.db.psqlCause
 import evaka.core.shared.domain.BadRequest
+import evaka.core.shared.domain.Conflict
 import evaka.core.shared.domain.EvakaClock
 import evaka.core.shared.domain.FiniteDateRange
 import evaka.core.shared.domain.NotFound
@@ -77,6 +85,8 @@ import evaka.core.shared.security.Action
 import evaka.core.shared.security.PilotFeature
 import evaka.core.shared.security.actionrule.AccessControlFilter
 import java.time.LocalDate
+import org.jdbi.v3.core.statement.UnableToExecuteStatementException
+import org.postgresql.util.PSQLState
 import org.springframework.format.annotation.DateTimeFormat
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
@@ -90,10 +100,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/employee/daycares")
-class DaycareController(
-    private val daycareService: DaycareService,
-    private val accessControl: AccessControl,
-) {
+class DaycareController(private val accessControl: AccessControl) {
     @GetMapping
     fun getDaycares(
         db: Database,
@@ -289,7 +296,9 @@ class DaycareController(
                         Action.Unit.READ_GROUPS,
                         daycareId,
                     )
-                    daycareService.getDaycareGroups(it, daycareId, from, to, includeClosed)
+                    if (!it.isValidDaycareId(daycareId))
+                        throw NotFound("No daycare found with id $daycareId")
+                    it.getDaycareGroups(daycareId, from, to, includeClosed)
                 }
             }
             .also {
@@ -309,22 +318,24 @@ class DaycareController(
         @RequestBody body: CreateGroupRequest,
     ): DaycareGroup {
         return db.connect { dbc ->
-                dbc.transaction {
+                dbc.transaction { tx ->
                     accessControl.requirePermissionFor(
-                        it,
+                        tx,
                         user,
                         clock,
                         Action.Unit.CREATE_GROUP,
                         daycareId,
                     )
-                    daycareService.createGroup(
-                        it,
-                        daycareId,
-                        body.name,
-                        body.startDate,
-                        body.initialCaretakers,
-                        body.aromiCustomerId,
-                    )
+                    tx.createDaycareGroup(
+                            daycareId,
+                            body.name,
+                            body.startDate,
+                            body.aromiCustomerId,
+                        )
+                        .also {
+                            tx.initCaretakers(it.id, it.startDate, body.initialCaretakers)
+                            tx.createDaycareGroupMessageAccount(it.id)
+                        }
                 }
             }
             .also { group ->
@@ -395,9 +406,23 @@ class DaycareController(
         @PathVariable groupId: GroupId,
     ) {
         db.connect { dbc ->
-            dbc.transaction {
-                accessControl.requirePermissionFor(it, user, clock, Action.Group.DELETE, groupId)
-                daycareService.deleteGroup(it, groupId)
+            dbc.transaction { tx ->
+                accessControl.requirePermissionFor(tx, user, clock, Action.Group.DELETE, groupId)
+                try {
+                    if (tx.hasGroupPlacements(groupId))
+                        throw Conflict("Cannot delete group which has children placed in it")
+                    tx.deleteDaycareGroupMessageAccount(groupId)
+                    tx.deleteDaycareGroup(groupId)
+                } catch (e: UnableToExecuteStatementException) {
+                    throw e.psqlCause()
+                        ?.takeIf { it.sqlState == PSQLState.FOREIGN_KEY_VIOLATION.state }
+                        ?.let {
+                            Conflict(
+                                "Cannot delete group which is still referred to from other data",
+                                cause = e,
+                            )
+                        } ?: e
+                }
             }
         }
         Audit.UnitGroupsDelete.log(targetId = AuditId(groupId))
